requirements.txt
requirements.txt
python-telegram-bot[ext]==21.0.1
Flask==3.0.3
gunicorn==22.0.0
requests==2.32.3
PyPDF2==3.0.1
